requirements.txt
Flask>=2.3
Jinja2>=3.1
Werkzeug>=2.3
MarkupSafe>=2.1
markdown>=3.4.1
peewee==3.13.3
itsdangerous==2.1.2
Click==8.1.3
python-dotenv==0.10.5
gunicorn==20.1.0
